+++ 1556750042560-DouSuroMigration.ts
@@ -1,0 +1,11 @@
+import {MigrationInterface, QueryRunner} from "typeorm";
+
+export class DouSuroMigration1556750042560 implements MigrationInterface {
+
+    public async up(queryRunner: QueryRunner): Promise<any> {
+    }
+
+    public async down(queryRunner: QueryRunner): Promise<any> {
+    }
+
+}
