+++ backend/src/controllers/UserManager.ts
@@ -1,4 +1,4 @@
-import { User } from "../models";
+import { User } from "../entity";
 
 export class UserManager {
   static async createAccount(email: string, password: string): Promise<User> {
