+++ backend/src/middleware/authMiddleware.ts
@@ -1,7 +1,7 @@
 import { Response, NextFunction } from "express";
 import { AuthRequest } from "../types/AuthRequest";
 import { SecretsService } from "../controllers/SecretsService";
-import { User } from "../models";
+import { User } from "../entity";
 
 export default async function authenticate(
   req: AuthRequest,
