+++ backend/src/types/AuthRequest.ts
@@ -1,5 +1,5 @@
 import { Request } from "express";
-import { User } from "../models";
+import { User } from "../entity";
 
 export interface AuthRequest extends Request {
   user?: User;
